refactor: Move the code merging write infos to generated_types to share
